+++ persistence/src/main/scala/com/codefactory/ecommerce/repositories/CartRepository.scala
@@ -16,8 +16,7 @@ case class CartRepository() extends LazyLogging with QueryVariable {
 //                 db: backend.Database) = db.run {
 //    (carts.map(c => (c.productID, c.quantity))
 //      returning carts.map(_.id)
-//      into ((productQ, id) =>
-//              Cart(id, productQ._1, productQ._2)) += (product, quantity))
+//      into ((productQ, id) => Cart(id, productQ._1, productQ._2)) += (product, quantity))
 //  }
 
   def putCart()(implicit ec: ExecutionContext,
@@ -30,20 +29,4 @@ case class CartRepository() extends LazyLogging with QueryVariable {
        .update(toUpdate.quantity, toUpdate.productID))
   }
 
-  // Needs to be changed to post only product Id and quantity
-  //   val quantityComparisonOk = for {
-  //      (p, c) <- products join carts on (_.cartProductID === _.id)
-  //      if p.quantity >= c.quantity
-  //    } yield c.status
-  //    db.run(quantityComparisonOk.result)
-  //
-  //    val quantityComparisonNotOk = for {
-  //      (p, c) <- products join carts on (_.cartProductID === _.id)
-  //      if p.quantity < c.quantity
-  //    } yield c.status
-  //    db.run(quantityComparisonNotOk.result) -NO LONGER APPLICABLE KEPT AS A PLACEHOLDER
-//  case Some(hotel) => {
-//    val updatedHotel = Hotel(hotel.id, toUpdate.name, toUpdate.address, toUpdate.zip)
-//    db.run(hotels.filter(_.id === id).update(updatedHotel)).map(_ => Some(updatedHotel))
-//  }
 }